+++ app/src/main/res/layout/pregamepage.xml
@@ -48,13 +48,37 @@
                 android:layout_gravity="fill_horizontal"/>
 
             <TextView
+                android:text="Start Platform: "
+                android:textSize="@dimen/text_medium"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:paddingVertical="15dp"
+                android:paddingRight="50dp"
+                android:layout_alignTop="@+id/tableLayout"
+                android:layout_row="1"
+                android:layout_column="0"
+                android:layout_columnSpan="1"
+                />
+
+            <Spinner
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textSize = "@dimen/text_small"
+                android:id="@+id/autoStartPlatform"
+                android:spinnerMode="dropdown"
+                android:layout_row="1"
+                android:layout_column="1"
+                android:layout_columnSpan="2"
+                android:layout_gravity="fill_horizontal"/>
+
+            <TextView
                 android:text="Started with:  "
                 android:textSize="@dimen/text_medium"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:paddingVertical="15dp"
                 android:layout_alignTop="@+id/tableLayout"
-                android:layout_row="1"
+                android:layout_row="2"
                 android:layout_column="0"
                 android:layout_columnSpan="1" />
 
@@ -65,7 +89,7 @@
                 android:id="@+id/startingObjectsCargo"
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
-                android:layout_row="1"
+                android:layout_row="2"
                 android:layout_column="1"
                 android:checked="false"
                 android:text="Cargo"
@@ -74,7 +98,7 @@
             <CheckBox
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
-                android:layout_row="1"
+                android:layout_row="2"
                 android:layout_column="2"
                 android:id="@+id/startingObjectsHatch"
                 android:layout_columnWeight="1"
